[adding-payment-api-endpoint] Renamed contributionsFrontendService to PaymentApiService
